requirements.txt
python-dotenv==0.21.0
langchain-core==0.3.76
langgraph==0.3.30
langchain-google-genai==2.1.10
langchain-community==0.3.20
pandas==2.0.0
pdfplumber==0.11.7
streamlit==1.49.1